Moved the download gui to new file
