+++ api.py
@@ -1,5 +1,5 @@
-from googleapiclient.discovery import build  # Importing build function from googleapiclient.discovery
-from abc import ABC, abstractmethod  # Importing ABC and abstractmethod for creating abstract classes
+from googleapiclient.discovery import build
+from abc import ABC, abstractmethod
 
 
 # Abstract class for API commands
