+++ last_used_directory.txt
@@ -1,0 +1,1 @@
+C:/Users/alexa/Documents/GitHub/music-player/music
