+++ main_app.py
@@ -1,25 +1,14 @@
-from customtkinter import CTkButton, CTkEntry
-from tkinter import Tk, Toplevel, Label, END, messagebox, Listbox
+from customtkinter import CTkButton
+from tkinter import Tk, filedialog
 from song_menu import create_gui, get_songs
-from api import search
-import subprocess
+from download_page import download_song, get_last_used_directory, set_last_used_directory
 from abc import ABC, abstractmethod
 
 
-# Function to download audio from YouTube using yt-dlp
-def download_audio(url, path):
-    print(f"Starting download for {url}...")
-    subprocess.run([
-        'yt-dlp',
-        '-f', 'bestaudio',
-        '-o', f'{path}/%(title)s.%(ext)s',
-        '--extract-audio',
-        '--audio-format', 'mp3',
-        '--add-metadata',
-        url
-    ])
-
-
+# TODO: add queue and skip feature
+# TODO: add live radio feature for Lab2
+# TODO: better GUI
+# TODO: something else I forgot
 # Abstract class for observers
 class Observer(ABC):
     @abstractmethod
@@ -28,7 +17,7 @@ class Observer(ABC):
 
 
 # Observable class for implementing observer pattern
-class Observable: #TODO notify the user when a song is downloaded
+class Observable:
     def __init__(self):
         self._observers = []
 
@@ -50,115 +39,64 @@ class ButtonCommand(ABC):
         pass
 
 
-class DownloadSongCommand(ButtonCommand):
-    def execute(self):
-        download_song()
-
-
 class PlaySongCommand(ButtonCommand):
     def execute(self):
         play_song()
 
 
 class ExitCommand(ButtonCommand):
+    def __init__(self, root):
+        self.root = root
+
     def execute(self):
-        root.quit()
+        self.root.quit()
 
 
-# Factory class for creating button commands
 class ButtonFactory:
     @staticmethod
-    def create_button(type):
+    def create_button(type, root=None):
         if type == "Download":
-            return DownloadSongCommand()
+            return DownloadSongCommand(root)
         elif type == "Play":
             return PlaySongCommand()
         elif type == "Exit":
-            return ExitCommand()
+            return ExitCommand(root)
         else:
             raise ValueError("Invalid button type")
 
 
-# Decorator function to print start and end of download
-def download_decorator(func): #TODO: remove this soon
-    def wrapper(*args, **kwargs):
-        print("Starting download...")
-        result = func(*args, **kwargs)
-        print("Download finished!")
-        return result
-
-    return wrapper
-
-
-# Function to download a song
-@download_decorator
-def download_song():
-    # Initialize variables and GUI elements
-    results = []
-    listbox = Listbox()
-
-    # Function to handle song selection from listbox
-    def on_song_select(event):
-        print("Song selected...")
-        nonlocal results
-        widget = event.widget
-        selection = widget.curselection()
-        video_id = results[selection[0]]['videoId']
-        print(f"Selected video ID: {video_id}")
-        video_url = f"https://www.youtube.com/watch?v={video_id}"
-        print(f"Video URL: {video_url}")
-        download_path = r"C:\Users\alexa\Documents\GitHub\music-player\music"
-        print(f"Download path: {download_path}")
-        download_audio(video_url, download_path)
-        download_window.destroy()
-        messagebox.showinfo("Download Complete", "The song has been downloaded successfully!")
-
-    # Function to submit search query and display results
-    def submit(event=None):
-        nonlocal results
-        nonlocal listbox
-        query = entry.get()
-        results = search(query)
-        listbox.delete(0, END)
-        for result in results[:5]:
-            listbox.insert(END, result['title'])
-
-    # Creating GUI for downloading a song
-    download_window = Toplevel(root)
-    download_window.geometry("500x500")
-    Label(download_window, text="Enter a song title: ").pack()
-    entry = CTkEntry(download_window, width=200, height=25)
-    entry.place(x=0, y=0)
-    entry.pack()
-    entry.bind('<Return>', submit)
-    listbox = Listbox(download_window, width=60)
-    listbox.pack()
-    listbox.bind('<<ListboxSelect>>', on_song_select)
-    CTkButton(download_window, text="Submit", command=submit).pack()
-
-
 # Function to play a downloaded song
-def play_song():  # TODO: implement flexible directory instead of hardcoded
-    directory = r"C:\Users\alexa\Documents\GitHub\music-player\music"
+def play_song():
+    directory = filedialog.askdirectory(initialdir=get_last_used_directory())
+    set_last_used_directory(directory)
     songs = get_songs(directory)
     create_gui(songs, directory)
 
 
 # Main GUI class that acts as an observer
 class MainGui(Observer):
+    def __init__(self):
+        self.root = Tk()
+        self.root.geometry("500x500")
+
     def update(self, *args, **kwargs):
         pass
 
     # Function to create the main GUI
     def create_main_gui(self):
-        global root
-        root = Tk()
-        root.geometry("500x500")
-        CTkButton(root, text="Download new song from YouTube",
-                  command=ButtonFactory.create_button("Download").execute).pack()
-        CTkButton(root, text="Play a downloaded song", command=ButtonFactory.create_button("Play").execute).pack()
-        CTkButton(root, text="Exit", command=ButtonFactory.create_button("Exit").execute).pack()
-        root.mainloop()
+        CTkButton(self.root, text="Download new song from YouTube",
+                  command=ButtonFactory.create_button("Download", self.root).execute).pack()
+        CTkButton(self.root, text="Play a downloaded song", command=ButtonFactory.create_button("Play").execute).pack()
+        CTkButton(self.root, text="Exit", command=ButtonFactory.create_button("Exit", self.root).execute).pack()
+        self.root.mainloop()
+
+
+class DownloadSongCommand(ButtonCommand):
+    def __init__(self, root):
+        self.root = root
+
+    def execute(self):
+        download_song(self.root)
 
 
 # Main function to run the application
